Update feedback authorization functionality and fixbugs

+++ TasteTrailExperience/src/TasteTrailExperience.Infrastructure/Feedbacks/Services/FeedbackService.cs
@@ -2,9 +2,11 @@ using Microsoft.AspNetCore.Identity;
 using Npgsql.TypeMapping;
 using TasteTrailData.Core.Feedbacks.Models;
 using TasteTrailData.Core.Users.Models;
+using TasteTrailExperience.Core.Common.Exceptions;
 using TasteTrailExperience.Core.Feedbacks.Dtos;
 using TasteTrailExperience.Core.Feedbacks.Repositories;
 using TasteTrailExperience.Core.Feedbacks.Services;
+using TasteTrailExperience.Core.Venues.Repositories;
 
 namespace TasteTrailExperience.Infrastructure.Feedbacks.Services;
 
@@ -12,12 +14,15 @@ public class FeedbackService : IFeedbackService
 {
     private readonly IFeedbackRepository _feedbackRepository;
 
+    private readonly IVenueRepository _venueRepository;
+
     private readonly UserManager<User> _userManager;
 
-    public FeedbackService(IFeedbackRepository feedbackRepository, UserManager<User> userManager)
+    public FeedbackService(IFeedbackRepository feedbackRepository,  IVenueRepository venueRepository, UserManager<User> userManager)
     {
         _feedbackRepository = feedbackRepository;
         _userManager = userManager;
+        _venueRepository = venueRepository;
     }
 
     public async Task<List<FeedbackGetDto>> GetFeedbacksByCountAsync(int count)
@@ -75,22 +80,34 @@ public class FeedbackService : IFeedbackService
 
     public async Task<int> CreateFeedbackAsync(FeedbackCreateDto feedback, User user)
     {
+        var venue = await _venueRepository.GetByIdAsync(feedback.VenueId) ?? 
+            throw new ArgumentException($"Venue by ID: {feedback.VenueId} not found.");
+
         var newFeedback = new Feedback
         {
             Text = feedback.Text,
             Rating = feedback.Rating,
             CreationDate = DateTime.Now.ToUniversalTime(),
             UserId = user.Id,
-            VenueId = feedback.VenueId,
+            VenueId = venue.Id,
         };
 
         var feedbackId = await _feedbackRepository.CreateAsync(newFeedback);
-
+        await UpdateVenueRatingAsync(newFeedback.VenueId);
+        
         return feedbackId;
     }
 
-    public async Task<int?> DeleteFeedbackByIdAsync(int id)
+    public async Task<int?> DeleteFeedbackByIdAsync(int id, User user)
     {
+        var feedbackToUpdate = await _feedbackRepository.GetAsNoTrackingAsync(id);
+
+        if (feedbackToUpdate is null)
+            return null;
+
+        if (feedbackToUpdate.UserId != user.Id)
+            throw new ForbiddenAccessException();
+
         var feedbackId = await _feedbackRepository.DeleteByIdAsync(id);
 
         return feedbackId;
@@ -99,18 +116,39 @@ public class FeedbackService : IFeedbackService
 
     public async Task<int?> UpdateFeedbackAsync(FeedbackUpdateDto feedback, User user)
     {
+        var feedbackToUpdate = await _feedbackRepository.GetAsNoTrackingAsync(feedback.Id);
+
+        if (feedbackToUpdate is null)
+            return null;
+
+        if (feedbackToUpdate.UserId != user.Id)
+            throw new ForbiddenAccessException();
+
+        var venue = await _venueRepository.GetByIdAsync(feedbackToUpdate.VenueId);
+
         var updatedFeedback = new Feedback
         {
             Id = feedback.Id,
             Text = feedback.Text,
             Rating = feedback.Rating,
             CreationDate = DateTime.Now.ToUniversalTime(),
-            UserId = user.Id,
-            VenueId = feedback.VenueId,
+            UserId = user.Id
         };
 
         var feedbackId = await _feedbackRepository.PutAsync(updatedFeedback);
+        await UpdateVenueRatingAsync(venue!.Id);
 
         return feedbackId;
+    }
+
+    private async Task UpdateVenueRatingAsync(int venueId) {
+        var venue = await _venueRepository.GetByIdAsync(venueId);
+
+        if (venue is null)
+            throw new ArgumentNullException(nameof(venueId));
+
+        venue.OverallRating = (float)Math.Round(venue.Feedbacks.Average(f => f.Rating), 2);
+
+        await _venueRepository.PutAsync(venue);
     }
 }
